+++ srbuild/project/profile.py
@@ -5,6 +5,15 @@ from srbuild.project.target import Target
 from srbuild.graph.graph import Graph
 from typing import List, Union, Dict
 
+import os
+
+# Inserts suffix into path, just before the extension
+def _file_suffix(path: str, suffix: str, ext: str = None) -> str:
+    split = os.path.splitext(path)
+    basename = split[0]
+    ext = ext or split[1]
+    return f"{path}.{suffix}" + f".{ext}" if ext else ""
+
 # Each profile has a Graph for linked/compiled targets. The source tree (i.e. FileManager) is shared.
 # Profiles can have default properties that are applied to each target within.
 # TODO: Add compiler/linker as a property of Profile.
@@ -18,31 +27,30 @@ class Profile(object):
     # libs can contain either Nodes from this graph, or paths to libraries, or names of libraries
     # TODO(0): Convert Targets in libs to Nodes.
     # This cannot be called with a target_config that has not been
-    def target(self, name, source_nodes, flags, libs, compiler, include_dirs, linker, lib_dirs) -> LinkedNode:
-        flags = self.flags + target_config.flags
-        libs = target_config.libs
-        compiler = target_config.compiler
-        include_dirs = target_config.include_dirs
-        linker = target_config.linker
-        lib_dirs = target_config.lib_dirs
+    def target(self, basename, source_nodes, flags, libs: List[Union[Node, str]], compiler, include_dirs, linker, lib_dirs) -> LinkedNode:
+        # Per-target flags always overwrite profile flags.
+        flags = self.flags + flags
 
         # First, add or retrieve object nodes for each source.
         object_nodes = []
         for source_node in source_nodes:
             # Only the include dirs provided by the user are part of the hash. When the automatically deduced
             # include_dirs change, it means the file is stale, so name collisions don't matter (i.e. OK to overwrite.)
+            # TODO: Maybe push signature generation into Generator.
             obj_sig = compiler.signature(source_node.path, include_dirs, flags)
-            obj_basename = os.path.basename(os.path.splitext(source_node.path)[0])
-            obj_path = os.path.join(self.build_dir, f"{obj_basename}.{obj_sig}.o")
+            obj_path = os.path.join(self.build_dir, _file_suffix(source_node.path, obj_sig, "o"))
             # User defined includes are always prepended the ones deduced for SourceNodes.
-            object_nodes.append(CompiledNode(obj_path, source_node, compiler, include_dirs))
+            obj_node = CompiledNode(obj_path, source_node, compiler, include_dirs, flags)
+            object_nodes.append(self.graph.add(obj_node))
 
-        # For any libraries that are paths or Nodes, treat as inputs.
-        # For any libraries that are names, pass them along to the linker.
-        # TODO: FIXME: This should respect the order of libs
-        input_libs, libs = _process_libs(libs)
-        # TODO(0): Node adding for absolute paths should happen in Project, needs to be added to FileManager.
-        lib_nodes = [self.graph.add(Node(lib)) for lib in input_libs]
-
+        # For any libraries that are Nodes, add as inputs to the final LinkedNode.
+        # For any libraries that are names, pass them along to the linker as-is.
+        lib_nodes: List[Node] = [lib for lib in libs if isinstance(lib, Node)]
+        # Next, convert all libs to paths or names.
+        libs: List[str] = [lib if not isinstance(lib, Node) else lib.path for lib in libs]
         # Finally, add the actual linked node
-        return self._add_linked_node(object_nodes + lib_nodes, flags, linker, libs, lib_dirs)
+        input_paths = [node.path for node in object_nodes]
+        linked_sig = linker.signature(input_paths, libs, lib_dirs, flags)
+        linked_path = os.path.join(self.build_dir, _file_suffix(basename, linked_sig))
+        linked_node = LinkedNode(linked_path, object_nodes, linker, libs, lib_dirs, flags)
+        return self.graph.add(linked_node)