+++ srbuild/project/project.py
@@ -1,4 +1,4 @@
-from srbuild.graph.node import CompiledNode, LinkedNode
+from srbuild.graph.node import Node, CompiledNode, LinkedNode
 from srbuild.project.file_manager import FileManager
 from srbuild.project.profile import Profile
 from srbuild.tools import compiler, linker
@@ -6,6 +6,7 @@ from srbuild.tools.flags import BuildFlags
 from srbuild.project.target import Target
 from srbuild.graph.graph import Graph
 from srbuild.logger import G_LOGGER
+
 from typing import List, Set, Union, Dict, Tuple
 from collections import OrderedDict
 import inspect
@@ -31,13 +32,55 @@ class Project(object):
         self.profile(name="release", flags=BuildFlags().O(3).std(17).march("native").fpic())
         self.profile(name="debug", flags=BuildFlags().O(0).std(17).debug().fpic())
 
-    def _get_source_nodes(self, sources) -> List[CompiledNode]:
+
+    def _target(self, basename: str, sources: List[str], flags: BuildFlags, libs: List[Union[Target, str]], compiler: compiler.Compiler, include_dirs: List[str], linker: linker.Linker, lib_dirs: List[str]) -> Target:
+
         # Convert sources to full paths
-        source_nodes: List[CompiledNode] = [self.files.source(path) for path in sources]
-        G_LOGGER.verbose(f"For sources: {sources}, found source paths: {source_nodes}")
-        return source_nodes
+        def get_source_nodes(sources: List[str]) -> List[CompiledNode]:
+            source_nodes: List[CompiledNode] = [self.files.source(path) for path in sources]
+            G_LOGGER.verbose(f"For sources: {sources}, found source paths: {source_nodes}")
+            return source_nodes
+
+        # The linker expects libs to be either absolute paths, or library names.
+        # e.g. ["stdc++", "/path/to/libtest.so"]
+        # If the library is provided as a path, we also add it as a node to the file manager
+        # so that we can properly rebuild when it is updated (even if it's external).
+        def get_libraries(libs: List[Union[Target, str]]) -> List[Union[Target, Node, str]]:
+
+            # Determines whether lib looks like a path, or like a library name.
+            def is_lib_path(lib: str) -> bool:
+                has_path_components = os.path.sep in lib
+                has_ext = bool(os.path.splitext(lib)[1])
+                return has_path_components or has_ext
+
+            fixed_libs = []
+            for lib in libs:
+                # Targets are handled by each profile individually
+                if not isinstance(lib, Target):
+                    candidates = self.files.find(lib)
+                    if is_lib_path(lib):
+                        if len(candidates) > 1:
+                            G_LOGGER.warning(f"For library: {lib}, found multiple candidates: {candidates}. Using {candidates[0]}. If this is incorrect, please provide a longer path to disambiguate.")
+                        # Add the library to the file manager as an external path
+                        lib = self.files.external(lib)
+                    elif candidates:
+                        G_LOGGER.warning(f"For library: {lib}, found matching paths: {candidates}. However, {lib} appears to be a library name rather than a path to a library. If you meant to use the path, please provide a longer path to disambiguate.")
+                fixed_libs.append(lib)
+            G_LOGGER.debug(f"Using fixed libs: {fixed_libs}")
+            return fixed_libs
+
+        source_nodes = get_source_nodes(sources)
+        libs: List[Union[Target, Node, str]] = get_libraries(libs)
+        target = Target()
+        for profile_name, profile in self.profiles.items():
+            # Process targets so we only give each profile its own LinkedNodes.
+            # Purposely don't convert all libs to paths here, so that each profile can set up dependencies correctly.
+            target_libs = [lib if not isinstance(lib, Target) else lib[profile_name] for lib in libs]
+            target[profile_name] = profile.target(basename, source_nodes, flags, libs, compiler, include_dirs, linker, lib_dirs)
+        return target
 
     # TODO: Docstrings
+    # Both of these functions will modify name before passing it to profile so that the filename is correct.
     def executable(self,
                     name: str,
                     sources: List[str],
@@ -47,10 +90,7 @@ class Project(object):
                     include_dirs: List[str] = [],
                     linker: linker.Linker = linker.clang,
                     lib_dirs: List[str] = []) -> Target:
-        source_nodes = self._get_source_nodes(sources)
-        self.executables[name] = Target()
-        for profile_name, profile in self.profiles.items():
-            self.executables[name][profile_name] = profile.target(name, source_nodes, flags, libs, compiler, include_dirs, linker, lib_dirs)
+        self.executables[name] = self._target(linker.exec_basename(name), sources, flags, libs, compiler, include_dirs, linker, lib_dirs)
         return self.executables[name]
 
     def library(self,
@@ -62,10 +102,7 @@ class Project(object):
                 include_dirs: List[str] = [],
                 linker: linker.Linker = linker.clang,
                 lib_dirs: List[str] = []) -> Target:
-        source_nodes = self._get_source_nodes(sources)
-        self.libraries[name] = Target()
-        for profile_name, profile in self.profiles.items():
-            self.libraries[name][profile_name] = profile.target(name, source_nodes, flags + BuildFlags().shared(), libs, compiler, include_dirs, linker, lib_dirs)
+        self.libraries[name] = self._target(linker.lib_basename(name), sources, flags + BuildFlags().shared(), libs, compiler, include_dirs, linker, lib_dirs)
         return self.libraries[name]
 
     # Before configuring, populate the source field of None