+++ srbuild/project/target.py
@@ -1,0 +1,6 @@
+from srbuild.graph.node import LinkedNode
+from typing import NewType, Dict
+
+# A Dict[str, LinkedNode] that maps profile names to the LinkedNodes for a target
+class Target(dict):
+    pass