+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Dates are in YYYY-MM-DD format.
 - Split `FileManager`'s '`source_info` into `source`, which adds sources to the graph, and `scan` which scans for include directories. Removed `includes` since that information is now part of the `Node`.
 - `FileManager` now tracks root directory and build directory instead of `Project` tracking it directly.
 - Smarter `find` function in `FileManager`. Additionally, `source` uses `find` to make sure sources exist.
+- Added `external` to `FileManager` to be able to track files external to the project.
+- `FileManager`'s `find` will now accept absolute paths that are outside the project.
+- `Profile` now has runnable `target` function (untested).
+- Improved handling of libraries in `Project`
 
 ## SRBuild v0.1.0 (2019-02-16)
 - Initial version, basic compiler functionality.